backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
pydantic-settings
pydantic[email]
pgvector
libonvif
opencv-python-headless
numpy
# faiss-gpu-cu12 ships GPU-enabled FAISS for CUDA 12 on PyPI.
# It replaces faiss-cpu; do NOT install both simultaneously.
faiss-gpu-cu12
onnxruntime-gpu==1.20.1
scikit-learn
confluent-kafka
fastavro
attrs
gdown
authlib
cachetools
minio
redis[asyncio]
pillow
ultralytics
insightface
scipy
httpx
python-multipart
websockets
bcrypt
python-jose[cryptography]
pyotp
qrcode
python-dotenv
psutil
prometheus_client
